Fix: TownRepository Long으로 변경

+++ src/main/kotlin/com/example/seollyongdanbackend/repository/TownRepository.java
@@ -3,5 +3,5 @@ package com.example.seollyongdanbackend.repository;
 import com.example.seollyongdanbackend.entity.Town;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TownRepository extends JpaRepository<Town, Integer> {
+public interface TownRepository extends JpaRepository<Town, Long> {
 }
